Add whitelisting of RW api endpoints
events and groups endpoints now only accept GET requests
All POST, PUT and PATCH must now go through the all_events
and all_groups endpoints.

A new whitelist variable can now be used in the ansible conf.
It must be a comma separated list of IP addresses or CIDR
ranges. If it is defined, access to the RW endpoints will be
limited to these IP.

Note that 127.0.0.1 will be accepted anyway.